refactor(client): add current location button and worker daily appointments screen
- Introduced a new `CurrentLocationButton` component for obtaining the user's
  current location and handling permissions.
- Added `WorkerDailyAppointmentsScreen` to display daily appointments for
  workers, including a map view and list view.
- Updated navigation to include the new services screen and adjusted imports
  for better organization.
- Refactored category components to improve structure and maintainability.

This update enhances user experience by providing location-based services
and a more organized appointment management interface.
