+++ composeApp/src/commonMain/kotlin/com/example/seviya/app/navigation/AppNavGraph.kt
@@ -19,7 +19,6 @@ import androidx.navigation.compose.composable
 import androidx.navigation.toRoute
 import com.example.seviya.feature.shared.ProfessionalProfileScreen
 import com.example.seviya.feature.shared.RequestAppointmentScreen
-import com.example.seviya.feature.landing.ServicesScreen
 import com.example.seviya.feature.landing.RoleAdmissionCatalogScreen
 import com.example.seviya.feature.landing.RoleCatalogScreen
 import com.example.seviya.feature.client.navigation.clientNavGraph
@@ -35,22 +34,16 @@ import com.example.seviya.core.designsystem.components.WorkerTab
 import com.example.seviya.core.navigation.CategoriesCatalog
 import com.example.seviya.core.navigation.ClientAgenda
 import com.example.seviya.core.navigation.ClientAlerts
-import com.example.seviya.core.navigation.ClientAppointmentDetail
 import com.example.seviya.core.navigation.ClientConfiguration
-import com.example.seviya.core.navigation.ClientDailyAppointments
 import com.example.seviya.core.navigation.ClientDashboard
 import com.example.seviya.core.navigation.ClientFavorites
 import com.example.seviya.core.navigation.ClientHome
-import com.example.seviya.core.navigation.ClientLocationCatalog
 import com.example.seviya.core.navigation.ClientMap
 import com.example.seviya.core.navigation.ClientMessages
-import com.example.seviya.core.navigation.ClientPaymentUpload
 import com.example.seviya.core.navigation.ClientProfile
 import com.example.seviya.core.navigation.ClientRequests
 import com.example.seviya.core.navigation.ClientSearch
 import com.example.seviya.core.navigation.ClientSettings
-import com.example.seviya.core.navigation.ClientToWorkerReview
-import com.example.seviya.core.navigation.ClientWeeklyAppointments
 import com.example.seviya.core.navigation.Landing
 import com.example.seviya.core.navigation.ProfessionalProfile
 import com.example.seviya.core.navigation.RequestAppointment
@@ -62,6 +55,7 @@ import com.example.seviya.core.navigation.WorkerDashboard
 import com.example.seviya.core.navigation.WorkersList
 import com.example.seviya.feature.categories.CategoriesCatalogScreen
 import com.example.seviya.feature.landing.LandingScreen
+import com.example.seviya.feature.landing.ServicesScreen
 import com.example.shared.domain.entity.Appointment
 import com.example.shared.presentation.calendar.MonthlyCalendarViewModel
 import com.example.shared.presentation.requestAppointment.RequestAppointmentDraft
