+++ composeApp/src/commonMain/kotlin/com/example/seviya/feature/categories/components/CategoriesCatalogContinueBar.kt
@@ -1,4 +1,4 @@
-package com.example.seviya.UI
+package com.example.seviya.feature.categories.components
 
 import androidx.compose.foundation.BorderStroke
 import androidx.compose.foundation.layout.fillMaxWidth
