+++ composeApp/src/commonMain/kotlin/com/example/seviya/feature/categories/components/CategoriesCatalogHeader.kt
@@ -1,4 +1,4 @@
-package com.example.seviya.UI
+package com.example.seviya.feature.categories.components
 
 import androidx.compose.animation.core.RepeatMode
 import androidx.compose.animation.core.animateFloat
