+++ composeApp/src/commonMain/kotlin/com/example/seviya/feature/categories/components/CategoriesCatalogMenu.kt
@@ -1,4 +1,4 @@
-package com.example.seviya.UI
+package com.example.seviya.feature.categories.components
 
 import androidx.compose.animation.AnimatedVisibility
 import androidx.compose.animation.fadeIn
