+++ composeApp/src/commonMain/kotlin/com/example/seviya/feature/categories/components/CategoriesCatalogVisuals.kt
@@ -1,4 +1,4 @@
-package com.example.seviya.UI
+package com.example.seviya.feature.categories.components
 
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.vector.ImageVector
