+++ composeApp/src/commonMain/kotlin/com/example/seviya/feature/worker/WorkerRequestsScreen.kt
@@ -1,4 +1,4 @@
-package com.example.seviya.ui
+package com.example.seviya.feature.worker
 
 import androidx.compose.animation.AnimatedVisibility
 import androidx.compose.animation.core.FastOutSlowInEasing
@@ -40,6 +40,7 @@ import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
